backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1

# LLM & AI
langchain==0.1.0
langchain-community==0.0.10
langgraph==0.0.20
openai==1.3.7
dashscope==1.17.0  # Qwen API

# Vector Database
pymilvus==2.3.4

# Knowledge Graph
neo4j==5.14.1

# Document Processing
pypdf2==3.0.1
python-docx==1.1.0
pdfplumber==0.10.3

# Embeddings
sentence-transformers==2.2.2

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

